Mover lógica para a store
